Factor out build string helpers
